SELinux: check OPEN on truncate calls
Patchwork-id: 26950
O-Subject: [PATCH RHEL6 -v2] SELinux: check OPEN on truncate calls
Bugzilla: 578841
RH-Acked-by: James Morris <jmorris@redhat.com>

In BZ 578841 we realized that the SELinux sandbox program was allowed to
truncate files outside of the sandbox.  The reason is because sandbox
confinement is determined almost entirely by the 'open' permission.  The idea
was that if the sandbox was unable to open() files it would be unable to do
harm to those files.  This turns out to be false in light of syscalls like
truncate() and chmod() which don't require a previous open() call.  Upstream
I have proposed a new permission (or rework of the open permission) to actually
be checked on the last component of pathname resolution, but it is clearly
unacceptable for RHEL6 due to the invasive nature, both in kernel and in the
userspace policy changes that will be required.  Instead I looked at the syscalls
that did not have an associated 'open' check and found that truncate(), unlike
the others, did not have a seperate permission and even if it did have a
separate permission such a permission owuld be inadequate for use by sandbox
(since it owuld have to be granted so liberally as to be useless).  My RHEL6
solution is to check 'open' on truncate calls.  My upstream solution is yet
to be completed.

RH-Signed-off-by: Eric Paris <eparis@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>