zfcp: Register SCSI devices after fc_remote_port_add
Patchwork-id: 26998
O-Subject: [RHEL6 PATCH 1/6] [s390x] zfcp: Register SCSI devices after
	fc_remote_port_add
Bugzilla: 612586

Description
-----------
After a remote port failure, SCSI device files are missing.

When the ELS ADISC returns, it schedules a call to fc_remote_port_add,
but it does not trigger the registration of SCSI devices.
If this is the last recovery after a remote port failure and the SCSI
devices have been deleted from the dev_loss_tmo, the step to reregister
the SCSI devices is missing.

Change the registration of SCSI devices to be triggered
after the fc_remote_port_add call. For the initial inquiry
command to succeed the unit must also be open. If the
unit reopen is still pending, the inquiry command to the
LUN will be deferred with DID_IMM_RETRY, so there is no
harm from this approach.

Bugzilla
--------
BZ 612586
https://bugzilla.redhat.com/show_bug.cgi?id=612586

Upstream status of the patch
----------------------------
The patch has been sent upstream for inclusion into a 2.6.35-rc level.
http://marc.info/?l=linux-scsi&m=127857608417692&w=2

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>