zfcp: Use forced_reopen in terminate_rport_io callback
Patchwork-id: 27001
O-Subject: [RHEL6 PATCH 2/6] [s390x] zfcp: Use forced_reopen in
	terminate_rport_io callback
Bugzilla: 612621

Description
-----------
When using the fast_io_fail_tmo in non-NPIV mode, the
rport is set to online, although the actual port is not
available in the SAN.

When the port connection is shared with another operating
system in non-NPIV mode, the other OS can keep the port
connection open, and thus the reopen can succeed when the
port is not in the SAN.

Use the forced_reopen operation to force a logout. The
following "open" has to login to the remote port and
can only succeed when the remote port is active in the SAN.

Bugzilla
--------
BZ 612621
https://bugzilla.redhat.com/show_bug.cgi?id=612621

Upstream status of the patch
----------------------------
The patch has been sent upstream for inclusion into a 2.6.35-rc level.
http://marc.info/?l=linux-scsi&m=127857608317686&w=2

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>