cfq: always return false from should_idle if slice_idle is set to zero
Patchwork-id: 27013
O-Subject: [RHEL 6 PATCH] cfq: always return false from should_idle if
	slice_idle is set to zero
Bugzilla: 616904
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

Hi,

In testing a competing fsync-ing process and a sequential reader on
mid-grade storage, I found that cfq was incapable of achieving the I/O
rates of deadline, no matter how it was tuned.  Investigation, and insight
from Vivek (mostly the latter), led to identifying that we were still
idling for the last queue in the service tree.

Modifying cfq_should_idle to not idle when slice_idle is set to zero got
us much closer to the performance of deadline for this workload.

This is not yet upstream, though Jens promised to get this into 2.6.35.
This addresses bugzilla 616904.

Cheers,
Jeff

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>