Log msg when getting Unit Attention
Patchwork-id: 27039
O-Subject: [PATCH RHEL 6]: scsi: Log msg when getting Unit Attention
Bugzilla: 585432
RH-Acked-by: David Milburn <dmilburn@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This if for bz 585432.

If the user accidentally changes LUN mappings or it occurs
due to bug in the target, then it can cause data corruption that can take
months and months to track down. This patch adds a log
message when getting REPORT_LUNS_DATA_CHANGED and it adds
a generic message for other Unit Attentions with asc == 0x3f.

This is in the scsi tree here:
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=a953809f9d1deb97bb27716d773cadd2feebccf8

I tested it by adding/removing luns or changing the size.