ACPI: Export EDID blocks to the kernel
Patchwork-id: 27046
O-Subject: [RHEL6 PATCH 1/2] ACPI: Export EDID blocks to the kernel
Bugzilla: 616860
RH-Acked-by: Dave Airlie <airlied@redhat.com>

From: Matthew Garrett <mjg@redhat.com>

The ACPI spec includes a provision for hardware to provide EDID via the
ACPI video extension. In the KMS world it's necessary for a way to obtain
this from within the kernel. Add a function that either returns the EDID
for the provided ACPI display ID or the first display of the provided type.
Also add support for ensuring that devices with legacy IDs are supported.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Acked-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Len Brown <len.brown@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>