nouveau: fix race condition when under memory pressure
Patchwork-id: 27047
O-Subject: [RHEL6 PATCH] drm/nouveau: fix race condition when under memory
	pressure
Bugzilla: 602663
RH-Acked-by: Dave Airlie <airlied@redhat.com>

rhbz#602663

When VRAM is running out it's possible that the client's push buffers get
evicted to main memory.  When they're validated back in, the GPU may
be used for the copy back to VRAM, but the existing synchronisation code
only deals with inter-channel sync, not sync between PFIFO and PGRAPH on
the same channel.  This leads to PFIFO fetching from command buffers that
haven't quite been copied by PGRAPH yet.

This patch marks push buffers as so, and forces any GPU-assisted buffer
moves to be done on a different channel, which triggers the correct
synchronisation to happen before we submit them.

After discussion with another nouveau developer, it was agreed that while
this patch is fine in itself, that we'd prefer to work out a nicer, but
likely much more invasive, fix upstream.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>