ksmd and khugepaged freezing
Patchwork-id: 27048
O-Subject: [RHEL6 PATCH] ksmd and khugepaged freezing
Bugzilla: 617430
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Chris Wright <chrisw@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>

kswapd does this, so even if I think freezer should be smarter than
this when dealing with lowlatency schedule friendly kernel threads
like this, it's safer to comply with kswapd.

https://bugzilla.redhat.com/show_bug.cgi?id=617430

This is a forward port from:

http://harpoon.virt.bos.redhat.com/~ehabkost/kvm-patches/kvm-KSM-suspend-fix.patch

This is a build including the below patch, if somebody can test
suspend to ram and suspend to disk with this it'd be helpful.

https://brewweb.devel.redhat.com/taskinfo?taskID=2619614

=====
Subject: freeze khugepaged and ksmd

From: Andrea Arcangeli <aarcange@redhat.com>

It's unclear why schedule friendly kernel threads can't be taken away by the
CPU through the scheduler itself. It's safer to stop them as they can trigger
memory allocation, if kswapd also freezes itself to avoid generating I/O they
have too.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>