nfsd: fix error handling in __write_ports_addxprt
Patchwork-id: 27057
O-Subject: [RHEL6 PATCH 3/7] BZ#599675: nfsd: fix error handling in
	__write_ports_addxprt
Bugzilla: 599675
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

__write_ports_addxprt calls nfsd_create_serv. That increases the
refcount of nfsd_serv (which is tracked in sv_nrthreads). The service
only decrements the thread count on error, not on success like
__write_ports_addfd does, so using this interface leaves the nfsd
thread count high.

Fix this by having this function call svc_destroy() on error to release
the reference (and possibly to tear down the service) and simply
decrement the refcount without tearing down the service on success.

This makes the sv_threads handling work basically the same in both
__write_ports_addxprt and __write_ports_addfd.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>