nfsd: just keep single lockd reference for nfsd
Patchwork-id: 27059
O-Subject: [RHEL6 PATCH 5/7] BZ#599675: nfsd: just keep single lockd reference
	for nfsd
Bugzilla: 599675
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Right now, nfsd keeps a lockd reference for each socket that it has
open. This is unnecessary and complicates the error handling on
startup and shutdown. Change it to just do a lockd_up when starting
the first nfsd thread just do a single lockd_down when taking down the
last nfsd thread. Because of the strange way the sv_count is handled
this requires an extra flag to tell whether the nfsd_serv holds a
reference for lockd or not.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>