Fix device cgroup not allowing access to a partition
Patchwork-id: 27077
O-Subject: [RHEL6 PATCH] cgroup: Fix device cgroup not allowing access to a
	partition
Bugzilla: 589662
RH-Acked-by: Chris Wright <chrisw@redhat.com>

BZ: 589662

cgroup device controller would not allow access to a device partition until
and unless one allowed access to whole device also. This happened because
while one tries to open a partition, __blkdev_get() also tries to do
__blkdev_get() on whole device containting partition so that partition
device bdev can point to whole device block_device.

Fix this by not checking for device access permissions if we are doing
__blkdev_get() on whole device due to a partition being opened.

Chris Wright posted following patch upstream. It is in Andrew's tree and
might make into Linus tree either in 2.6.35, otherwise in 2.6.36.

I have done some testing with this patch and it looks good to me.

The cgroup device whitelist code gets confused when trying to grant
permission to a disk partition that is not currently open.  Part of
blkdev_open() includes __blkdev_get() on the whole disk.  This simply
avoids the cgroup check for the whole disk case when opening a
partition.

https://bugzilla.redhat.com/show_bug.cgi?id=589662

Signed-off-by: Chris Wright <chrisw@sous-sol.org>
Acked-by: Serge E. Hallyn <serue@us.ibm.com>
Tested-by: Serge E. Hallyn <serue@us.ibm.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>