add context argument to shrinker callback
Patchwork-id: 27086
O-Subject: [RHEL6 PATCH 1/3] mm: add context argument to shrinker callback
Bugzilla: 617035
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

Upstream commits: 7f8275d0d660c146de6ee3017e1e2e594c49e820
		  567c7b0edec0200c5c6613f07c3d3b4034fdc836
BZ: 617035

The current shrinker implementation requires the registered callback
to have global state to work from. This makes it difficult to shrink
caches that are not global (e.g. per-filesystem caches). Pass the shrinker
structure to the callback so that users can embed the shrinker structure
in the context the shrinker needs to operate on and get back to it in the
callback via container_of().

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>