cpufreq: pcc driver should check for pcch method before calling _OSC
Patchwork-id: 27130
O-Subject: [RHEL 6.0 PATCH] cpufreq: pcc driver should check for pcch method
	before calling _OSC
Bugzilla: 616908
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

bz: #616908

The pcc specification documents an _OSC method that's incompatible with the
one defined as part of the ACPI spec. This shouldn't be a problem as both
are supposed to be guarded with a UUID. Unfortunately approximately nobody
(including HP, who wrote this spec) properly check the UUID on entry to the
_OSC call. Right now this could result in surprising behaviour if the pcc
driver performs an _OSC call on a machine that doesn't implement the pcc
specification. Check whether the PCCH method exists first in order to reduce
this probability.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>