ehca: Catch failing ioremap()
Patchwork-id: 27157
O-Subject: [PATCH RHEL6.0 BZ617747] ib/ehca: Catch failing ioremap()
Bugzilla: 617747
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Matthew Garrett <mjg@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=617747

Description:
============
ehca catch ioremap fail

While allocating resources (queue pairs and completion queues) in our driver,
we use the ioremap() function to establish a mapping in virtual memory to the
physical memory where the hardware resource can be accessed. The driver does
not handle errors (= NULL pointer) which may be returned by ioremap(), and
accesses the NULL memory later on, which leads to a NULL pointer dereference.

Fix this behavior by checking for NULL pointers returned by ioremap and return
an error to the calling function.

During the testing for HPC, we ran into the situation where ioremap() returned
NULL and the system crashed when trying to access this mapped memory.

When ioremap() fails with a NULL pointer, catch the error and pass it to the
caller of create_qp() or create_cq() instead of trying to dereference the NULL
pointer later on.

Signed-off-by: Alexander Schmidt <alexs@linux.vnet.ibm.com>