return EINVAL when thawing unfrozen filesystems
Patchwork-id: 27162
O-Subject: [PATCH RHEL6] fs: return EINVAL when thawing unfrozen filesystems
Bugzilla: 601324
RH-Acked-by: Josef Bacik <josef@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

This is for bug 601324 -
echo j > /proc/sysrq-trigger causes infinite unfreeze/Thaw event

I just sent the following trivial patch upstream to resolve
the issue:

The iterate_supers() loop in do_thaw_all() assumes
that we can call thaw_bdev() until we hit an error; when
I wrote this, we got -EINVAL when requesting a thaw of
an unfrozen bdev.  (it called until error because we
may have nested freezes).

However, commit 4504230a (freeze_bdev: grab active
reference to frozen superblocks) changed this so that
we get a return of 0 in this case, and do_thaw_all()
spins forever when an emergency system-wide thaw
is requested.

It looks to me like the change was inadvertent,
so this patch changes it back.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>