drop matroxfb by default
Patchwork-id: 27184
O-Subject: [PATCH] [RHEL6.0 PATCH] drop matroxfb by default
Bugzilla: 604830
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

bz 604830
old brew build: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2568015

IBM are attempting to fix this upstream, apparantly remote console works,
local monitor doesn't. However the server seems to work fine in text mode
and X starts okay, I'm not sure what the requirement for a graphical console is
and I think its getting a bit late to fix this for EL6.

Some notes:
In EL5 we had this enabled by default, but we never added support for the new
G200 chips so we never seen any issues. Now in EL6 I've attempted to install on
a new server with the G200eV in it and when matroxfb loads it loses sync.

Now matroxfb is required on Power systems for a graphics console, but on x86
we have no requirement to load this driver, and the X driver seems to work fine
here. The powerpc config files have the matrox bit enabled explicitly.

Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>