GFS2: Backup plan for "vmalloc is slow"
Patchwork-id: 27187
O-Subject: [RHEL6] GFS2: Backup plan for "vmalloc is slow" (bz #619331)
Bugzilla: 619331
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This patch, taken from upstream, is designed to reduce the dependence of
GFS2 upon vmalloc. It uses kmalloc whenever the size requested is small
enough that this is a possible solution and thus vmalloc will only now
be used for the very largest directories.

This is for bz #619331

It does not solve the problem of vmalloc being slow, but instead this is
a backup plan to try and avoid the consequences in GFS2 specifically.
I'd still like to see vmalloc performance improved, although there has
been some discussion upstream that leads me to think that might not
happen, at least in the near future.

This has been tested for correctness with various different sizes of
directory, but has yet to be tested for performance. It is though,
relatively low risk, and at the very worst, will not make things any
slower.

Even if vmalloc is eventually improved, this patch will still make sense
to do. This patch also has the advantage of using GFP_NOFS rather than
the implied GFP_KERNEL of the original vmalloc version. That hasn't
actually caused a problem so far, but its safer to use GFP_NOFS at this
point in the code.

Steve.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>