e1000e: 82577/82578 PHY register access issues
Patchwork-id: 27201
O-Subject: [PATCH RHEL6] e1000e: 82577/82578 PHY register access issues
Bugzilla: 592480
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

During device resets, PHY register read errors are occurring on the
82577 device.  These errors are random, but when they occur the device
is left in an unstable state.  A driver reload and sometimes even a
reboot necessary to restore functionality.

Generally, a short series of resets is enough to cause the following
error:

kernel: 0000:00:19.0: eth0: Error reading PHY register.

The following script can be used to reproduce the problem:

while true
do
	ethtool -t $1
	if (dmesg | grep "Error reading PHY")
	then
		echo "PHY Error reported!"
		exit 1
	fi
	sleep 1
done

This was recently resolved upstream and a backport of this this patch
should do it:

	commit ff847ac2d3e90edd94674c28bade25ae1e6a2e49
	Author: Bruce Allan <bruce.w.allan@intel.com>
	Date:   Tue, 27 Jul 2010 12:28:46 +0000

	    e1000e: 82577/82578 PHY register access issues

	    The MAC-PHY interconnect on 82577/82578 uses a power management feature
	    (called K1) which must be disabled when in 1Gbps due to a hardware issue on
	    these parts.  The #define bit setting used to enable/disable K1 is
	    incorrect and can cause PHY register accesses to stop working altogether
	    until the next device reset.  This patch sets the register correctly.

This will resolve 592480.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>