page allocator: Drain per-cpu lists after direct reclaim allocation fails
Patchwork-id: 27261
O-Subject: [RHEL6 PATCH] Drain per-cpu lists after direct reclaim allocation
	fails
Bugzilla: 614427
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=614427
https://bugzilla.redhat.com/attachment.cgi?id=432045
=====
Subject: Drain per-cpu lists after direct reclaim allocation fails

From: Mel Gorman <mel@csn.ul.ie>

When under significant memory pressure, a process enters direct reclaim
and immediately afterwards tries to allocate a page. If it fails and no
further progress is made, it's possible the system will go OOM.

However, one systems with large amounts of memory, it's possible that a
significant number of pages are on per-cpu lists and inaccessible to the
calling process. This leads to a process entering direct reclaim more
often than it should increasing the pressure on the system and
compounding the problem.

This patch notes that if direct reclaim is making progress but
allocatioons are still failing that the system is already under heavy
pressure. In this case, it drains the per-cpu lists and tries the
allocation a second time before continuing.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>