Default Xen PV-HVM to off
Patchwork-id: 27267
O-Subject: [RHEL6 PATCH] Default Xen PV-HVM to off
Bugzilla: 618172
RH-Acked-by: Andrew Jones <drjones@redhat.com>

BZ 618172

This patch puts the Xen PV-HVM support back to disabled by
default in RHEL6, and puts back the original beta-1 'xen_pv_hvm=enable'
kernel cmdline parameter if it is to be used in a RHEL6 xen HVM guest.

Problem:
Xen fully virtualized/HVM installations of RHEL6 (on RHEL5-dom0)
fail due to network initialization by NetworkManager, which anaconda
is reliant on.
This is new behavior/dependency of anaconda in RHEL6, as this
type of installation works with RHEL4 & RHEL5 with xen pv-hvm
configured into those kernels.

The base problem comes from the fact that the RHEL5 xen-hvm tools
present two networking devices with the same mac address to a xen HVM
guest: 8139 & xen-vnif. This 'presentation' was hidden in RHEL6 until
xen pv-hvm support was built-in to the RHEL6 kernel (snap-8).
NetworkManager is choking on two devices with
the same mac address, and fails to configure a network, so network-based
installations fail (common case install method for virtualization guests).

This network failure is not seen if a pre-rhel6-xen-pvhvm configuration
is upgraded with a kernel with xen-pvhvm, which was how the support
was tested by engineering.
In fact, taking a rhel6 snap8(beta2-refresh) PV installation (guest disk image)
and using it as a RHEL6 HVM guest (by creating a xen hvm guest configuration
file that uses the same image) boots successfully and configures
the xen-vnif. So, this problem appears to be subtle with the use &/or
timing of NM in the installation path.

This (unique) problem with RHEL6 installation may require some
major surgery to NM.  Since an upgrade to the RHEL5.6 qemu is planned
to effectively disable/disconnect the 8139 on RHEL6 xen-pv-hvm(enabled)
guests, this problem has a near-term solution but dependent on rhel5 upgrade.
That solution also requires kernel cmdline parameters to enable 8139 disconnect.
Thus, adding  a kernel cmdline to enable the use of xen pv-hvm drivers on
RHEL6 xen FV guests is a similar (xen) setup requirement.

This patch (actually a more complicated one that included smp enabling)
was included in the Beta-1 RHEL6 kernel.  This (much simpler) patch puts it back in place.

Please review and ack.

- Don

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>