megaraid: fix sas expander issue
Patchwork-id: 27284
O-Subject: [RHEL6 PATCH] megaraid: fix sas expander issue
Bugzilla: 607930
RH-Acked-by: Mike Christie <mchristi@redhat.com>

Resolves bz#607930

If a disk is pulled (a member of a redundant RAID array such
as RAID1 or even an unconfigured disk) the system is getting a panic.
The patch comes from LSI and has been tested by customer - FJ.
It is a small subset of the 4.31 update it is posted for RHEL5.
LSI is going to repost the upstream patch because of
bisectability problems which is no issue here.
We are not taking the whole driver update but only
a minimal patch to not cause another regression.

Brew - taskID=2645560

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>