avoid stalling allocations by introducing watermark_wait
Patchwork-id: 27285
O-Subject: [RHEL6 PATCH] mm: avoid stalling allocations by introducing
	watermark_wait
Bugzilla: 589604
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Waiting for IO in the page allocator results in very slow performance
when there is no IO going on.  This turns out to be a common scenario
on systems with lots of IO but fairly low memory, eg. virtual machines.
It was originally spotted on s390, but I imagine some of our KVM using
customers will run into it as well.

The patch is not upstream, because upstream has little interest in
machines with little memory and lots of IO.  I believe this kind of
setup will become more common with virtualization and I will try to
discuss this patch next week at the VM summit.

I believe we should merge the patch in RHEL6 because:

1) the patch fixes the bug in a self contained and obvious way

2) the patch does not interfere with upstream efforts in the
   pageout and writeback code, so it should not interfere with
   what we want/need to do in the VM for RHEL 6.1

3) this patch, or something like it, is likely to go upstream
   since "reducing page reclaim related latency" is a big theme
   upstream right now

Fixes bug 589604

------
Subject: [PATCH] mm: avoid stalling allocations by introducing watermark_wait

From: Mel Gorman <mel@csn.ul.ie>

Update - in addition to mels last patch this also contains:
- add a watermark_check_zone at the end of free_one_page according to comments
  from nick piggin
- rebase to match RHEL6 B1 sources

Under heavy memory pressure, the page allocator may call congestion_wait()
to wait for IO congestion to clear or a timeout. This is not as sensible
a choice as it first appears. There is no guarantee that BLK_RW_ASYNC is
even congested as the pressure could have been due to a large number of
SYNC reads and the allocator waits for the entire timeout, possibly uselessly.

At the point of congestion_wait(), the allocator is struggling to get the
pages it needs and it should back off. This patch puts the allocator to sleep
on a zone->pressure_wq for either a timeout or until a direct reclaimer or
kswapd brings the zone over the low watermark, whichever happens first.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>