bonding: allow arp_ip_targets on separate vlans to use arp validation
Patchwork-id: 27359
O-Subject: Re: [PATCH RHEL6 1/2] bonding: allow arp_ip_targets on separate
	vlans to use arp validation
Bugzilla: 581657
RH-Acked-by: David S. Miller <davem@redhat.com>

I added support upstream (and in RHEL5) for arp validation to work on
VLAN interfaces as well as the base bonding device.  Apparently
customers have found this helpful enough to request it on RHEL6 as well.

This is a backport of the following upstream commits:

	commit 1f3c8804acba841b5573b953f5560d2683d2db0d
	Author: Andy Gospodarek <andy@greyhouse.net>
	Date:   Mon Dec 14 10:48:58 2009 +0000

	    bonding: allow arp_ip_targets on separate vlans to use arp validation

	commit ca8d9ea30bc79b2965a1d169dcb2f48f02af4d2d
	Author: Andy Gospodarek <andy@greyhouse.net>
	Date:   Wed Jan 6 12:56:37 2010 +0000

	    fix bonding: allow arp_ip_targets on separate vlans to use arp validation

This will resolve RHBZ 581657.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>