fix unsupported hardware to only be power5
Patchwork-id: 27360
O-Subject: [PATCH RHEL6.0 BZ619507] powerpc: fix unsupported hardware to only be
	power5
Bugzilla: 619501
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=619507

Description:
============
only display hardware unsupported for power5. We are missing the
condition to determine if the system is power5. right now we display
the msg on all power systems, this patch fixes that.

Aristeu created the correct patch about a month ago and this patch
was missing when bz 600509 was checked in.

RHEL Version Found:
===================
RHEL 6.0

kABI Status:
============
No symbols were harmed.

Brew:
=====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2646165

Upstream:
=========
Doesn't apply to upstream, it is only a change for RHEL6.

Test Status:
============

I tested this on power6 system today and no msg was displayed. I didn't
have chance to test this on power5 today, but did test this about a month
ago when Aristeu created the patch.

===============================================================
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>