cifs: account for new creduid parameter in spnego upcall string
Patchwork-id: 27377
O-Subject: [RHEL6 PATCH 19/18] BZ#618608: cifs: account for new creduid=0x%x
	parameter in spnego upcall string
Bugzilla: 618608
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

The following problem was discovered by Bruce during review of this set.
I've tested it and pushed the fix upstream yesterday.

The commit that added the creduid=0x%x parameter failed to increase the
buffer allocation to account for it.

Reported-by: J. Bruce Fields <bfields@fieldses.org>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>